Add files via upload
I have no idea how to use github but I started working on the final. Added an introduction, made dummy variables, and put in a pairwise plot. The pairwise plot is kind of ugly and I think it might be better to split that up into a couple different pairwise plots or get rid of it all together